requirements.txt
boto3 == 1.*
coderedcms == 3.*
coderedcms-bootstrap4 == 2.*
django-cors-headers == 4.*
django-environ == 0.11.*
django-encrypted-model-fields == 0.6.*
django-health-check == 3.*
django-modelcluster == 6.*
django_sendmail_backend == 0.1.*
django-ses == 3.*
django-storages == 1.*
psycopg2 == 2.*
psutil == 5.*
wagtail == 5.*
# django-iam-dbauth
